[libromdata] RomDataFactory: Add a create() overload that takes a filename.
In order to thumbnail Wii U Packages, and probably some other packages
I'd like to add later on, we have to thumbnail a directory. This isn't
doable with IRpFile, so as a workaround, use a filename.

This overload will create an RpFile and pass it to the usual create()
function if it's a file. For directories, it currently does nothing,
but it will eventually do something.

doc/abi/: Add this function to the ABI lists.