[libwin32common] ComBase: Restore initFnPtrs() using pthread_once().
Removed the assertion, since the function pointers are now only
initialized on demand by isThemeActive().

Check for PTHREAD_ONCE_INIT before going into the loop. Otherwise, we
end up with an infinite loop.

The COM objects are loading properly with libromdata as a DLL, but the
thumbnail isn't showing up, and the "ROM Properties" tab is crsahing.