[win32] MessageWidget: Get the rect for the Close button in WM_SIZE.
Otherwise, it ends up getting negative coordinates, so it never shows
up and doesn't respond to mouse button presses.

I'm pretty sure this worked before. Not sure what changed...

This wasn't a big issue with RomDataView, but with KeyManagerTab, it
adjusts the position of the ListView, so it needs to be closable.