[libromdata] IsoPartition: Fix up some minor confusion between UTF-8 and cp1252.
Filenames are stored as cp1252 internally because that's how they're
stored on the disc, so this improves performance.

Filenames requested outside of IsoPartition are in UTF-8, so they have
to be converted to cp1252 in order to be compared.

(cherry picked from commit 84bf1de629aa956b2f62434a6a90229cb3ad0f6b)