c++11-compat.h: #define U8(x) as a u8"string" wrapper.
It's similar to _T() on Windows. Note that it only works on const strings.

NOTE: It works with character constants on gcc-12, but KDevelop is
complaining about it, which means clang doesn't support it yet.
https://reviews.llvm.org/D119221

Using static_cast<char8_t>() for char8_t constants for now.

NOTE 2: U8(RP_KDE_UPPER) doesn't seem to work, but RP_KDE_UPPER by
itself seems to be added to u8 strings just fine...