[win32] Pass POINT and SIZE by value instead of by const reference.
On i386, it eliminates an instruction, since it ends up passing the
entire SIZE on the stack.

On amd64, there's an extra shift instruction because it passes the SIZE as
a single 64-bit value in a register, but it eliminates memory accesses.

PVS-Studio V835 (Lo): Passing cheap-to-copy argument by reference may
lead to decreased performance.