[librptexture] ImageDecoder_Linear_sse2.cpp: Fix up some __m128i argument passing issues.
AppVeyor (MSVC 2015) failed to build on 32-bit due to the change from
`__m128i` to `const __m128i&`. It only failed on the fourth parameter,
so maybe I don't need to change all four, but since 32-bit isn't
guaranteed to have 16-byte stack alignment, I might as well change it.

Here's how it works now:
- If the compiler supports __vectorcall (MSVC 2013+), __vectorcall will
  be used and all __m128i parameters will be passed by value.
- Otherwise, if it's 64-bit, __m128i will be passed by value.
- Otherwise, if it's 32-bit, __m128i will be passed by const ref.

TODO: Optimize the mask setting to only be done when necessary?
Not sure if the compiler handles that by itself.