[libromdata] GdiReader: Don't use `volatile`.
This apparently breaks the range-based `for` loop:

V834 (Hi): Incorrect 'const volatile BlockRange&' type of the 'vbr' loop
variable. This leads to the variable binding to a temporary object
instead of a range element. The type of range elements is 'BlockRange'.

This fixes a regression from commit 9f3548a569bb61b1829293211886b512b9d1cc12.
(Use range-based `for` loops in more places.)