[win32] Use LoadLibraryEx() to harden library loading.
Restrict system DLLs to LOAD_LIBRARY_SEARCH_SYSTEM32.

For our own DLLs, we get the full path of the current module anyway,
so we don't need to add LOAD_LIBRARY_SEARCH_APPLICATION_DIR. This
wouldn't work in rom-properties.dll, since the application directory
would be C:\Windows\System32 (Explorer) or other directories.

On Windows XP, this flag is ignored, so there's no extra security,
but it doesn't break.

On Windows Vista and Windows 7, this flag requires KB2533623:
https://support.microsoft.com/kb/2533623

On Windows 8 and later, this flag is supported out-of-the-box.