[libromdata] GameCubeBNR: Rework string encoding handling.
Added a constructor that takes a gcnRegion parameter. This ensures that
we always have the correct encoding for BNR1 when opening.bnr is from a
GameCube disc image.

Added convenience functions to get each string and determine the
encoding based on gcnRegion (if present) and the presence of a cp1252
copyright symbol ('\xA9').

This fixes metadata display for, among other things, Pokémon Box (USA).