[libromdata] EXE_PE: Sort DLL imports by module, name, and hint.
NOTE: The std::sort() lambda function must return if the field being
compared is not equal, *not* if it's less than. It should return true
if less than and false otherwise, but it should not continue to
checking other fields. Otherwise, it'll get confused and end up going
out of bounds.