Revert "[librpfile] RpFile_stdio.cpp: Use open() with O_NONBLOCK."
This reverts commit d5587f3ba831781064d318522d9aad5d1a53ad7a.

We don't want to interfere with the contents of pipes or sockets,
since that could cause problems with programs that are using them.

The TOCTOU issue isn't too important here, since the "worst case"
is someone replaces a regular file with a pipe and causes fopen()
to hang.

TODO: Maybe do both: call statx()/stat() first, then open() with
O_NONBLOCK and a second statx()/fstat() to prevent TOCTOU.