appveyor.yml: Simplify the configuration without using matrix.
The matrix configuration wasn't working right, and I'm not sure how
to get it working with the addition of Ubuntu, so don't use it for
now.

The only downside is we won't be able to test MinGW-w64 builds, which
isn't that big of a deal since MinGW-w64 isn't a target platform at
the moment. (No support for isolation-aware contexts.)

NOTE: Re-enable VS2015 after testing this.