[libromdata] GameCubeSave: Trim trailing CRs.
TMNT Mutant Melee [GE5EA4] has CRs at the end of the filename and
description fields. I'm not sure if it's actually included as part
of the filename on the memory card, but it looks ugly here, so
remove the CR.

NOTE: rpcli was showing it as '␍', but UI frontends were sometimes
handling it as a linebreak.

TODO: Maybe make the UI frontends show it as '␍' and don't trim it?