Cleaned up GTK3/GTK4 LibNautilusExtension handling a bit.
Only one version can be installed at a time, so check once and then use
its path to calculate the other path for symlink purposes.

FIXME: Nautilus 43, the first GTK4 version, has a completely different
properties interface that only supports a key/value store and not
arbitrary GtkWidgets. We can probably make use of it for basic properties,
but not for tabs or lists. (Maybe bitfields, converted to text...)