[libromdata] EXE_NE: fix modref off-by-one error
Okay so it turns out this is 1-indexed, with absolutely no indication of
it in any of the format references I could find.

This also means that my previous assertion in commit message
c087b0c3de3a56ec589e8c29aecf2d30428e4a52 that "some modules don't have
names for the vast majority of [exported entries]" is probably
incorrect. I based that entirely on the fact that a lot of imports
wouldn't get resolved by EXENEEntries. In fact, when I switched it over
to Wine spec files, the situation didn't really improve, but at that
point I already turned off my brain, because clearly the .spec files are
the best source of data for EXENEEntries, so any unresolved imports must
mean it's a truly anonymous export, right? Well, .spec files have no way
to specify an export without a name, so in hindsight, an imported
anonymous entry is a good indicator that something is amiss (such an
import would never resolve against the Wine implementation of the DLL).