[librpbase] RomFields::Field: Change `name` from std::string to const char*.
std::string takes up 32 bytes on 64-bit Linux, and the field name is
considered read-only once it's added to Field, so change it to a regular
C string. This string is strdup()'d by the various RomFields functions
and is free()'d in RomFieldsPrivate::delete_data().

This reduces the size of RomFields::Field from 96 bytes to 72 bytes on
64-bit Linux.