[gtk] stub-export.c: Use a mostly-blank argv[] for GTK3/GTK4.
Otherwise, it complains and exits if other arguments, e.g. "-d", are
passed to rp-stub/rp-config, since we're not setting up command line
options in GApplication.

NOTE: Not doing this for the KDE or Windows frontends. For KDE, Qt will
handle its own options (if added after "--"), but it won't fail if it
doesn't recognize anything. Windows won't do anything by itself with
the command line arguments, unlike GTK and Qt.