[win32] AboutTab: Add an update check.
UpdateChecker is a C++ object but not a Win32 control. It sends custom
window messages to the specified HWND. Error messages and/or the retrieved
update version are stored in the C++ object, since the window message
can't easily transmit this data. (Well, it can, but I don't want to deal
with struct marshalling.)

The version label is shrunken a bit to make room for the update label.

TODO: Make it threaded.