NETWORK.md: Mention metered/unmetered networks.
Windows 10 supports indicating metered networks since RTM (2015), but
only through Windows Runtime. Windows 10 v2004 added a standard WinAPI
method, GetNetworkConnectivityHint(), which is what we'll be targetting.

If there's an easy way to implement support for the WinRT version without
breaking support for older versions of Windows, I can switch over to that
later on.

On Linux, NetworkManager's org.freedesktop.NetworkManager D-Bus interface
has a Metered property, which indicates if the active connection is
metered (or is guessed to be metered).