[kf5] rp-convert-to-png.desktop: Added a service menu to make it easy to convert supported image formats to PNG.
This uses a new option in rp-stub, '-a' / '--autoext', which creates the
output filename by changing the file extension of the source file.

TODO: Test on KDE4. If it works there with no or minimal changes, move
the .desktop file to a KDE common package. For now, it's in the KF5
package.