[win32] RP_ContextMenu: Use a separate thread to convert images to PNG.
d->filenames is now a pointer to std::vector<LPTSTR>. This pointer is
transferred to the conversion thread, which deletes it after it's done.

TODO: This stops Explorer from blocking, but now there's no visible
feedback if it takes a while. Maybe add a progress dialog (to all UI
frontends) or something...