[gtk] Use g_clear_handle_id() to stop timers.
This is similar to g_clear_object() and g_clear_pointer().

g_clear_handle_id() was added in glib-2.56, so it's now added to
glib-compat.h. The macro is defined regardless of version, since we
set the GLib minimum version to 2.34.0 if not targetting GTK4. We
redefined the macro without GLIB_AVAILABLE_MACRO_IN_2_56 so it doesn't
print warnings all the time when building the GTK2 and GTK3 plugins.