[librpcpu] cpuflags_x86.c: Use __cpuid() if it's available in cpuid.h.
gcc's __cpuid() macro has a workaround for some Winchip i386 CPUs not
properly zeroing %ebx and %ecx on i386. It's easier to simply use
gcc's __cpuid() macro directly instead of implementing the workaround
here.

This is similar to MSVC where we use the __cpuid() intrinsic if compiling
with MSVC 2005 or later.