[rp-download] rp-download.cpp: Fix check_newer 304 handling if not using verbose output.
Previously, it wouldn't check for a 304, so it'd just return EXIT_FAILURE
because it received no data. The `verbose` check has been reworked such
that this functions properly if verbose output is not enabled.

This broke the update checker in cases where the cached version file
matches the one on the server. (It'd show "ERROR" instead of the actual
latest version number.)