[debian] rules: Explicitly disable GTK4 for now.
Otherwise, the minimum GLib version is set to 2.66.0, which doesn't work
for most of the older Ubuntu distributions.

FIXME: Fix this such that if GLib 2.66.0 isn't found, but 2.34.0 or later
is found, it builds correctly and GTK4 is automatically disabled.

(cherry picked from commit c3efad1368370780871795e5050fa32f20a5730a)