[kf5,kf6] Install the plugin into namespaced directories if supported. (KF5 5.89 and later)
Don't install the .desktop files in this case.

ThumbCreator: Install a symlink. (TODO: ThumbnailCreator for KF6?)

TODO: Should we split it into two libraries? That might complicate things
with rp-stub, so I'll keep it as a single library with a symlink for now.

[libunixcommon] dll-search.c: Search the namespaced directories for KF5
and KF6 in addition to the base directories.