[kde] *.json.in, *.json: Remove "X-KDE-Protocol".
For some reason, "X-KDE-Protocol" was taking precedence over
"X-KDE-Protocols", which broke desktop:/.

Not changing the old .desktop files, since those seemed to work fine
previously...

NOTE: KF5 5.47.0's release notes [1] say that support for the
"X-KDE-Protocols" key was added to KPropertiesDialogPlugin in 5.47.0.
JSON support was added in KF5 5.89.0, so this shouldn't be a problem.

[1] https://kde.org/announcements/frameworks/5/5.47.0/