[librptexture] DirectDrawSurface: New private function getRGBFormatTableEntry().
This is basically most of what getPixelFormatName() was, except it
returns a pointer to the RGB_Format_Table_t* entry instead of the
string description.

getPixelFormatName() is now an inline function that calls
getPixelFormatName() and returns the string description if the
entry was found.

updatePixelFormat(): Call getRGBFormatTableEntry() to get the RGB format
table entry, then do the extra processing that was done here before.

Code size differences: (64-bit Gentoo Linux, gcc-12.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
  10150     752       0   10902    2a96 DirectDrawSurface.cpp.o [before]
   9958     752       0   10710    29d6 DirectDrawSurface.cpp.o [after]
   -192       0       0    -192     -c0 Difference