[librptexture] GodotSTEX: Fix v4 format flags.
We were copying the format flags from the pixel format value, which is
correct for v3 but not for v4. (v4 moves flags to another field)

Fix this by renaming pixelFormat_flags to format_flags and only storing
the masked-off format flags (v3) or the full format field (v4), and
initialize pixelFormat correctly at the same time.

This fixes format flags not showing up properly for v4 textures.