[librptexture] RpGdiplusBackend::shrink(): Save the new width and height values.
This might be what's causing the Sonic Colors JP ASTC test to fail.

The image is showing up as 480x272 on Windows. ASTC_8x8 uses 8x8 blocks,
so that matches the internal resolution; however, it's supposed to be
shrunken to the logical resolution, 480x270.

Not tested locally; will test this on AppVeyor.