[microtar] Cache the last tar header instead of reverse-seeking.
Reverse-seeking past the output buffer can take a while, and the time
increases exponentially the further we read into the file. The only
reverse seeking done by MicroTAR that causes this problem is seeking
back to the header after reading the file data.

Instead of doing that, cache the file header. This eliminates virtually
all reverse-seeking past the output buffer.

Simple benchmarks with debug builds:
- Linux (-Og): 5487 ms -> 2125 ms
- Windows (/Od): 48199 ms -> 39656 ms