[libromdata/tests] RomHeaderTest: Added SMS and GG SDSC header tests.
Needed to write a new program to zero out most of the ROM while leaving
the SDSC strings in place. This was made more complicated by the fact
that the SDSC strings can be anywhere within the first 64 KB of the ROM
(or potentially even after the first 64 KB if the string starts near the
end of it), but the SDSC header is always at 0x7FE0.

Before: (top-level)
- Lines: 10893/51922, 21.0%
- Functions: 745/3631, 20.5%

After: (top-level)
- Lines: 10943/51922, 21.1%
- Functions: 747/3631, 20.5%

NOTE: Two ROMs that were pulled out of the main SMS set due to containing
the string "SDSC" did not actually have an SDSC header; they merely
contain that string. These ROMs don't change code coverage at all, so
the SMS header package won't be rebuilt at the moment.

ROMs left out of the SMS (non-SDSC) header set:
- Damiana by Heliophobe (PD).sms
- Damiana by Heliophobe (PD) [a1].sms