[librptext] conversion.cpp: Initialize the localized decimal point using pthread_once(); handle LC_ALL=C.
If LC_ALL=C, always use '.' as the decimal point.

FIXME: If using C locale, don't do localization for the whole part. This
is handled on Linux, but not Windows. None of the RomHeaderTest test
suites currently depend on localized file sizes or frequencies, so this
isn't a problem yet.