[win32] RpImageWin32.cpp, toHBITMAP_mask(): Fix shadowed 'x' variables.
This may have caused the icon masks to be broken.

I'm not sure what the effects of this will be, so I'll have to verify
it on Windows later.

Note that icon masks are generally not used with 32-bit color icons,
so this will probably have minimal effect, if anything.

Found using cppcheck-2.10.