[win32] UpdateChecker::ThreadProc(): Close the version file.
Usually not an issue, since this is currently only called if the update
check is run in rp-config, but we should close the file anyway.

The GTK+ version is similar to the Windows version and has the fclose().
The Qt version uses QFile and has RAII, so this isn't an issue there.

Found using cppcheck-2.10.