[librpbase] RomFields, RomMetaData: Rework the copy/move constructors and assignment operator.
The copy-and-swap idiom seems to cause stack overflows when compiled using
MSVC 2015. Either that, or my implementation was busted.

Either way, the reworked version still uses copy-and-swap for operator=(),
but for the move constructor, it's directly coded similar to the copy
constructor, but different.

- Only the necessary data is copied.

- The old object has its type set to RFT_INVALID so the destructor won't
  attempt to delete anything.

- Neither function is inlined in the .hpp file anymore.

This should fix AppVeyor builds, since AppVeyor uses MSVC 2015.