[win32] LanguageComboBox: Revert to WNDCLASS instead of WNDCLASSEX.
For some reason, WNDCLASSEX was preventing it from functioning properly.
This resulted in the "Options" tab disappearing when it was clicked.

OptionsMenuButton: Not tested, but since it used the same setup as
LanguageComboBox, this has been reverted as well.

MessageWidget still uses WNDCLASSEX but it doesn't have this problem.

This fixes a regression from commit 86746b44000fc10d0ac5941ef534c5794d26237f.
([win32] AchWin32: Initialize wndClass using a static const struct.)