[librpbase] RpImageLoader: Make JPEG detection a bit more robust.
The first four bytes are *not* a magic number. The format spec says
the first byte is 0xFF, and the second byte is not 0xFF; the next
two bytes indicate the length of the "data bytes". It just so happened
that in all the JPEGs I tested, the data byte length was the same.

The `file` tool seems to indicate the third byte is 0xFF for JPEG,
and third+fourth are 0xFFF7 for JPEG-LS. Hence, we'll check that
the third byte is 0xFF as well.

References:
- https://en.wikipedia.org/wiki/JPEG_File_Interchange_Format
- https://github.com/file/file/blob/master/magic/Magdir/jpeg