[kde] LinuxAttrView: Port over the checkbox array stuff from GTK.
Instead of pre-defining checkboxes in the UI file, create them in a
`for` loop. Also re-translate them in a `for` loop.

This significantly reduces code size.

Code size difference: (64-bit Gentoo Linux, gcc-13.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
  19494       8       0   19502    4c2e LinuxAttrView.cpp.o [KF5] [before]
   7842     536       0    8378    20ba LinuxAttrView.cpp.o [KF5] [after]
 -11652    +528       0  -11124   -2b74 Difference

TODO: Move some of this to libromdata later. Doing so would require adding
new functions, which technically isn't an ABI break, but it would mean
that a newer version of rom-properties v2.2 would not work with an older
version of libromdata.so.3.