[gtk] DosAttrView: Use arrays to reduce code size a bit.
I'm not converting all of the checkboxes to an array right now, since
there's only six attributes. Instead, it's a union of both the struct
of named checkboxes and an array of checkboxes.

init() uses a loop to connect the signal.
update_attrs_display() uses a loop to update the checkboxes.

Code size difference: (64-bit Gentoo Linux, gcc-13.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
   3377       0      32    3409     d51 DosAttrView.c.o [GTK3] [before]
   3223       0      32    3255     cb7 DosAttrView.c.o [GTK3] [after]
   -154       0       0    -154     -9a Difference