[dbus] Disable unnecessary functions in interfaces in order to reduce the size of the generated wrapper code.
For NetworkManager, we only need the 'metered' property.
For Notifications, we only need the 'notify' method.

The full SpecializedThumbnailer1 interface is needed for
rp-dbus-thumbnailer, but it isn't used anywhere else.

.so size reductions: (64-bit Gentoo Linux, gcc-13.1.0, release build, no LTO)

rom-properties-xfce.so: 453,000 -> 395,656 (-57,344)
rom-properties-gtk3.so: 461,192 -> 403,848 (-57,344)
rom-properties-gtk4.so: 465,288 -> 407,944 (-57,344)
rom-properties-kf5.so:  694,680 -> 608,664 (-86,016)

.so size reductions: (64-bit Gentoo Linux, gcc-13.1.0, release build, LTO)

rom-properties-xfce.so: 375,176 -> 346,504 (-28,672)
rom-properties-gtk3.so: 387,464 -> 354,696 (-32,768)
rom-properties-gtk4.so: 313,736 -> 285,064 (-28,672)
rom-properties-kf5.so:  629,152 -> 559,520 (-69,632)

I'm still keeping the wrapper code because it means I don't have to write
boilerplate code for both QtDBus and GDBus. Granted, QtDBus is relatively
simple to use with direct access, but GDBus is more complicated.