[kde] RomDataView: Check for achievements in paintEvent(), not showEvent().
For some reason, we're getting a showEvent() when the propreties dialog is
opened in addition to when the tab is selected. paintEvent() only happens
when the tab is selected, so move achievement checking there.

Note that this doesn't happen in ConfigDialog; the update checkeer in
AboutTab only triggers when the tab is selected. It's probably a minor
implementation detail in KDE's KPropertiesDialog.