[librpsecure] restrict-dll.c: New file for restricting the DLL search path.
Useful for the standalone executables.
Not so much for the shell extensions.

This uses SetDefaultDllDirectories() if available. If not, it uses
SetDllDirectory() to at least remove CWD from the path.

Call rp_secure_restrict_directory_lookup() in:
- amiiboc
- rp-download
- rpcli
- svrplus
- rp-config

On Linux, this function does nothing, since Linux (and most other Unix)
systems don't loaded shared libraries from CWD in the first place.
(...unless someone adds "." to LD_LIBRARY_PATH for some reason...)