[libromdata/tests] RomHeaderTest: Add a few WIA/RVZ headers.
More than once, the WIA/RVZ parser has broken, resulting in a blank
properties tab. The last one I missed and ended up in v2.2, and an
issue was filed. (#389: Gamecube Property sheet is blank)

Adding this to RomHeaderTest should prevent this from happening by
causing AppVeyor to fail if it breaks again.

Currently, two disc images are included, in both WIA and RVZ formats,
with just 256 bytes from the header, and 4096 blank bytes to allow it
to pass tests in the GameCube constructor:

- rubysap-timerfix-1219-10-eng (GCN; "berry program update")
- swupdate_ALL_us (Wii; RVT IOS updater)

Code coverage changes:

Before:
  lines......: 22.3% (11572 of 52001 lines)
  functions..: 20.7% (734 of 3542 functions)

After:
  lines......: 22.4% (11673 of 52001 lines)
  functions..: 20.9% (740 of 3542 functions)