[libcachecommon] filterCacheKey(wchar_t*): Fix UTF-16 surrogate pair handling.
The first (high) code point must be U+D800-U+DBFF. The second (low)
code point must be U+DC00-U+DFFF. It seems my check was backwards.

In addition, the low surrogate check should be looking to see if
the second code point is *not* a low surrogate, instead of checking
if it *is* a high surrogate, since that potentially breaks if the
second code point isn't a surrogate pair code point at all.

TODO: Add unit tests for the wchar_t version.