[libcachecommon/tests] Add filterCacheKey(wchar_t*) tests.
NOTE: Runs on Windows only. (RP_WIS16)

NOTE 2: Implemented by converting UTF-8 to UTF-16. Because of this,
we have to skip the overlong UTF-8 sequence tests, so an extra parameter
is added that indicates if a test is usable for UTF-16 testing.

TODO: Add emoji tests and more non-BMP tests.

[libcachecommon] CacheKeys: Check for RP_WIS16 instead of _WIN32.
This is defined in config.librptext.h.