[libwin32common] DelayLoadHelper.c: Simplify the DLL list a bit.
Instead of two separate lists for NDEBUG and DEBUG, use a single list,
and add DEBUG_SUFFIX (which is "" for NDEBUG and "d" for DEBUG) to the
DLL names where necessary.

Moved dll_whitelist[] out of rp_loadLibrary().

TODO: Use strcpy_s() or similar to copy the DLL name in?