[win32] RP_ExtractIcon, RP_ExtractImage, RP_ShellPropSheetExt: Store Unicode (OLECHAR or TCHAR) filenames.
Windows allows arbitrary 16-bit code points without normalization, so it's
possible that a filename has an unpaired surrogate character that would
break when converting to/from UTF-8. To work around this, store the
filenames as the original UTF-16 version. (...or ANSI for ANSI builds?)

TODO: Need to add wchar_t* overloads for:
- FileSystem::isOnBadFS()
- RpFile constructor

Not doing that right not because it might need a DLL version bump, i.e.
romdata-3.dll -> romdata-4.dll.

Those functions still use UTF-8 filenames for now, which means they will
still have problems handling broken UTF-16 filenames.

For the IPersistFile-based classes (RP_ExtractIcon, RP_ExtractImage), the
filename is stored as LPOLESTR, which is WCHAR on both Unicode and ANSI
Windows. This requires some extra handling when converting to/from TCHAR.
Interestingly, this *does* allow some simplification for e.g. the
GetCurFile() function, since we're always using LPOLESTR, so no conversion
between UTF-8, UTF-16, or ANSI is needed.

For RP_ShellPropSheetExt, the filename is stored as LPTSTR.