[cmake] msvc.cmake: Consolidate "/guard:cf"; clean up some other checks.
The "/guard:cf" linker check was using the wrong variable, so it didn't
work correctly. Handle it in the compiler flag check instead.

Simplify the checks for "/CETCOMPAT".

Use two underscores in the "/Zc:throwingNew" check for consistency with
the other checks.

Only disable thread-safe statics on i386 for Windows XP compatibility.
We're supporting a minimum of Windows Vista for amd64, and for ARM CPUs,
the minimum is Windows 10.