[libromdata] DMGSpecialCases: Fix RomHeaderTest failures.
We need to trim the title before checking it. Otherwise, the extra spaces
prevent some games from being detected properly, e.g.:

Tom & Jerry (USA, Europe).gb

Also, make sure the title in the table has the same length as the title.
Otherwise, it's not a match. We need to do this because we're using
strncmp() with the non-NULL-terminated string directly from the ROM
header. Note that this check may be incorrect if the ROM's title is longer
than the title in the table, but in that case, the strncmp() will fail.

Also: "Primal Rage (USA, Europe) (Beta).gb" was correctly detected as
Japanese, but prior to DMGSpecialCases, the external URLs were missing
the "-J" specifier. I'm not sure *why* it was missing, but these commits
fixed that, so I had to rebuild the DMG.json.tar.zst and DMG.txt.tar.zst
files with the updated test results.

This should fix the AppVeyor build.