[win32] RP_ShellPropSheetExt: Truncate RFT_LISTDATA strings to 259+1 characters.
ListView is limited to 260 characters. The tcscpy_s() threw an assertion
if the source string was longer. Use tcsncpy_s() with _TRUNCATE instead,
whihc will truncate the string instead of asserting.

Also, when parsing the strings, if any of them are longer than 260
characters, truncate them and add "...".

Related to #392: WIM Support