[gtk3] Need to g_object_ref_sink() the XAttrView before g_object_unref()'ing it.
It has a floating reference at this point, so calling g_object_unref()
prints a warning:

(nautilus): Gtk-WARNING **: A floating object was finalized. This means
that someone called g_object_unref() on an object that had only a floating
reference; the initial floating reference is not owned by anyone
and must be removed with g_object_ref_sink().

This didn't show up on my local system since ext4 has extended attributes,
but VirtualBox's vboxsf does not.