[kde] Remove subdirectories from KDE #includes.
The subdirectories could either be present or not present in KF5 (and I
think KDE4), but with KF6, the subdirectories are themselves included
as part of KF6's CMake scripts. The top-level directory is *not*
included anymore.

Also, use the C++-style header names instead of .h for consistency.
(Should work on KDE4, KF5, and KF6.)