[gtk] KeyManagerTab_gtk3.cpp: Make sure tab->treeStore is unref()'d when tab->treeView is destroyed.
It's not exactly a memory leak, since rp-config is being closed at this
point, but we should handle it anyway.

Handling it here instead of in rp_key_manager_tab_dispose() because in
GTK3, tab->treeView is already destroyed, but in GTK4, it isn't.
Using a GWeakNotify function works in both cases.