[gtk] KeyManagerTab_gtk4.cpp: Load values and statuses.
Set the value column's label to use a monospace font.

The signal handlers are significantly easier to implement than with the
GTK4 version because we're modifying separate GListStores with objects
instead of a single giant GtkTreeStore with multiple values.

RpKeyStoreItem: Change "gboolean valid" to "uint8_t status" to match the
actual status value from KeyStoreUI. This also lets it show more than
just "valid" or "not valid".

TODO:
- Show the "Valid?" icon.
- GTK 4.12 adds GtkSectionModel, which is similar to what we have
  available on Windows.