[gtk] gtk-compat.h: gtk_css_provider_load_from_data() was deprecated in GTK 4.12.
A new function, gtk_css_provider_load_from_string(), is basically the
same, except it doesn't have a length argument; instead, it requires a
NULL-terminated string. We were always using a NULL-terminated string
with length -1 anyway, so this just requires some minor changes elsewhere
to use the new _STRING() macro and get rid of the length parameter.

NOTE: It may have been possible to use the length parameter with sizeof()
as a minor optimization, but it's not worth over-complicating the code
for that, especially since it's been deprecated.