[gtk] KeyManagerTab_gtk4.cpp: Print the GtkEditableLabel's text but don't try changing anything.
This seems to work without any garbage or dangling pointer issues.

It seems the issue is gtk_editable_get_text() returns the pointer to the
internal text buffer, but the value is getting set in KeyStoreGTK, which
emits a signal and causes the GtkEditableLabel's text to be updated again,
which results in a dangling pointer.