[libromdata/tests] RomHeaderTest: Regenerate test data for MMM01/MBC1M ROM headers.
For the .bin files, all of the ROM headers need to be included, and the
files have to be the full 1 MiB or 512 KiB. The non-header portions of
the files are zeroed out.

In addition, the first 256 bytes of each ROM header has been zeroed out.
This isn't actually part of the DMG header, which starts at 0x100.

Prior to updating the MMM01/MBC1M ROM headers, I tested a .tar.zst with
just the zeroing of 0x00-0xFF in the existing ROM headers.

Before: DMG.bin.tar.zst -> 491,486 bytes
After:  DMG.bin.tar.zst -> 241,739 bytes
Difference:               -249,747 bytes

A bit less than half the original size is a pretty good reduction.

With the additional MMM01/MBC1M ROM headers: 260,640 bytes (+18,901)

Also, it seems that 17 of the MMM01/MBC1M ROMs I tested were not present
in the DMG test suite files before, so that's 17 more tests.