[gtk] KeyManagerTab: Call gtk_window_present() on the GtkFileChooserDialog to make it visible.
I'm not sure when exactly this broke (possibly when it was rewritten to
use a signal handler instead of a modal event loop), but on GTK4, the
file chooser dialog no longer appeared when attempting to import a file.
(...actually, this may have been broken on GTK4 since GTK4 support was
added in v2.1, but it worked on GTK3 without gtk_window_present()...)

RomDataView_ops.cpp: Use gtk_window_present() instead of gtk_widget_show()
for the GtkFileChooserDialog, and call it on GTK4 in addition to GTK2/GTK3.
This hasn't been tested on GTK4 yet due to GTK4 Nautilus not supporting
full property pages, but I assume GtkFileChooserDialog would not have
shown up properly there, too.

TODO: Switch to GtkFileDialog on GTK 4.10 and later.