[win32] CMakeLists.txt: Restore "/dENABLE_DECRYPTION" for RCFLAGS.
This is still needed for IDD_CONFIG_KEYMANAGER in the localized
Win32 .rc files. It is *not* needed in libromdata.

As a result, the Key Manager tab wasn't showing up in the Windows build,
since the dialog resource was missing.

TODO: Move this to cmake/win32.cmake instead?

This fixes a regression from commit 09565732129fd06a425c2720bd30a6ea5ad4030c.
(Don't translate the KeyManager Import menu actions.)

Affects: 2.2.1