[librpfile] XAttrReader_posix.cpp: Fall back to XFS_IOC_FSGETXATTR if FS_IOC_FSGETXATTR isn't available.
They're identical, but the FS_ version was added for ext4 in kernel 4.5,
so older kernels won't have it defined; whereas the the XFS_ version has
likely been present since XFS was added to Linux in 2001.

If neither is available, then the functionality will be disabled.

Likewise for FS_IOC_GETFLAGS and EXT2_IOC_GETFLAGS.
FS_IOC_GETFLAGS was introduced sometime before Linux 3.0, but I can't
find when it was added right now. It's probably been around for a long
time, so it probably isn't *too* relevant.