[gtk] Initial XfsAttrView implementation for the xattrs tab.
Doing the GTK version first because I can't be bothered to deal with
Qt Designer at the moment.

NOTE: If the file does have XFS attributes, but xflags == 0 and
project_id == 0, XfsAttrView won't be displayed. This is because these
attributes are rarely used compared to others.

TODO: Show it anyway if the file system is XFS?

[librpfile] Add XfsAttrData.c and xfs_flags.h.

XAttrReader_posix.cpp: If neither FS_IOC_FSGETXATTR nor XFS_IOC_FSGETXATTR
are defined, define it along with fsxattr. On Ubuntu 16.04, these aren't
defined for some reason, but the ioctl() *does* work if using XFS, and
it shows the XFS xflags and project ID. (ext4 didn't support this ioctl
until kernel 4.5; Ubuntu 16.04 has 4.4.)