[librpfile] Ext2AttrData: Split the lsattr character out of the label string.
It's now stored in a separate char before the label string. Note that
this does waste space (3 bytes on 32-bit, 7 bytes on 64-bit), but it's
easier to add the character there in the .c file. Putting the character
at the end wouldn't have saved much space anyway, since the data is
stored in an array, so the pointers would be aligned anyway.

[locale] Removed the lsattr characters from the translated strings.

Code size difference: (64-bit Gentoo Linux, gcc-13.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
   1952     528       0    2480     9b0 Ext2AttrData.c.o [before]
   1886     704       0    2590     a1e Ext2AttrData.c.o [after]
    -66    +176       0    +110     +6e Difference