[libromdata] Wim: Add a wrapper function for XMLElement::IntText() for older versions of TinyXML2.
This function was added in TinyXML2 5.0.0.

TinyXML2 didn't have actual version number macros until 6.2.0, so we
have to use CHECK_CXX_SOURCE_COMPILES() in CMake to determine if
XMLElement::IntText() exists.

The wrapper function implements IntText() using QueryIntText() in the
same way later versions of TinyXML2 implement it if IntText() isn't
available in TinyXML2 itself.

This fixes the build on Ubuntu 16.04.