Change all "*.*" file filters to "*".
"*.*" is a DOS holdover that's still used on Windows for reasons, and
on Windows, I believe it can match files that *don't* have a file
extension, whereas that isn't the case on Linux.

[libwin32ui] WinUI: Convert "*" to "*.*" when displaying file dialogs.
Windows users are more familiar with "*.*". In my testing, both "*"
and "*.*" functioned the same on Windows, and both matched files with
no file extension.